fix fo rwindows

+++ lockfile_from_path.py
@@ -3,12 +3,16 @@ import shutil
 import sys
 from conans.test.utils.tools import NO_SETTINGS_PACKAGE_ID, TestClient
 
-me = os.path.abspath(os.path.dirname(__file__))
+
+me = os.path.abspath(os.path.dirname(__file__)).replace("\\", "/")
+
 
 if __name__ == '__main__':
     # Configure a test client to play with
     current_folder = os.path.join(me, '_working_dir')
-    shutil.rmtree(current_folder)
+    if os.path.exists(current_folder):
+        shutil.rmtree(current_folder)
+    os.makedirs(current_folder)
 
     t = TestClient(path_with_spaces=False, current_folder=current_folder)
     t.run("config set general.default_package_id_mode=recipe_revision_mode")
